% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexposuRe.R
\name{analyze_simulated_networks}
\alias{analyze_simulated_networks}
\title{Run full simulation}
\usage{
analyze_simulated_networks(
  n1,
  n2,
  n3,
  rho_min = 0,
  rho_max = 1,
  rho_inc = 0.1,
  a = 2,
  b = 1,
  N = 100,
  analyze,
  metric = "max",
  correct = T,
  plot_results = T
)
}
\arguments{
\item{n1}{the number of media outlets in the environment}

\item{n2}{the number of agents in the environment}

\item{n3}{the number of types of media outlets and agents in the environment. Must be less than n1 and n2.}

\item{rho_min}{the minimum value of the randomizing parameter to be used in the simulation. When rho is 0, agents strictly visit outlets of the same type as themselves. When rho is 1, agents visit outlets randomly.}

\item{rho_max}{the maximum value of the randomizing parameter to be used in the simulation. When rho is 0, agents strictly visit outlets of the same type as themselves. When rho is 1, agents visit outlets randomly.}

\item{rho_inc}{the value by which rho changes for each step of the simulation}

\item{a}{the exponent of the power-law distribution from which the reputation of the media outlets is drawn. Default is 2.}

\item{b}{the skewness parameter that determines how skewed the browsing behavior of the agents is. Can be negative or positive, but not 0. Default is 1.}

\item{N}{number of networks to build for each value of rho. Defaults to 100.}

\item{analyze}{what analysis to perform? Currently supports two: "cd" for community detection. "c" for centralization.}

\item{metric}{only applicable if analyze == "cd". Which variant of NMI to use for calculation. Default is "max". Can be "max", "min", "sqrt", "sum", "joint"}

\item{correct}{only applicable if analyze = "cd". If TRUE, NMI scores are scaled using the scaling factor. This corrects for overfitting by community detection algorithms. Default is TRUE.}

\item{plot_results}{if TRUE, a graph of the results is shown.}
}
\value{
a tibble with the results for each iteration within the simulation
}
\description{
Calls \code{simulate_single_network} with different values of rho and returns the results of the analysis of those selected networks
Two analyses are currently supported with support for more planned for the future. See the \code{analyze} parameter for details.
}
\examples{
res <- simulate_analyze_networks(n1 = 50, n2 = 30, n3 = 3, rho_min = 0, rho_max = 0.4, N = 5)
}
\references{
Mukerjee, S. (2021). A systematic comparison of community detection algorithms for measuring selective exposure in co-exposure networks. Scientific Reports, 11(1), 1-11.
}
