% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexposuRe.R
\name{get_simulated_network}
\alias{get_simulated_network}
\title{Get Simulated Network}
\usage{
get_simulated_network(
  n1,
  n2,
  n3,
  rho,
  a = 2,
  b = 1,
  show_network = FALSE,
  niter = 500
)
}
\arguments{
\item{n1}{The number of media outlets in the environment}

\item{n2}{The number of agents in the environment}

\item{n3}{The number of types of media outlets and agents in the environment. Must be less than n1 and n2.}

\item{rho}{The randomizing parameter. When rho is 0, agents strictly visit outlets of the same type as themselves. When rho is 1, agents visit outlets randomly.}

\item{a}{the exponent of the power-law distribution from which the reputation of the media outlets is drawn. Default is 2.}

\item{b}{the skewness parameter that determines how skewed the browsing behavior of the agents is. Can be negative or positive, but not 0. Default is 1.}

\item{show_network}{if TRUE, will plot the coexposure network. Default is FALSE.}

\item{niter}{used if \code{show_network} = TRUE. Passed to \code{plot.igraph} for visualizing network.}
}
\value{
\code{get_simulated_network} returns a list with three elements: \cr
\itemize{
\item \code{g} the coexposure network of media outlets \cr
\item \code{ag} the augmented coexposure network \cr
\item \code{outlet_dat} a tibble containing the details of the simulated media outlets
}
}
\description{
\code{get_simulated_network()} runs a single simulation and returns coexposure networks and additional information related to those networks. \cr
}
\examples{
res <- get_simulated_network(n1 = 50, n2 = 30, n3 = 3, rho = 0.1, show_network = TRUE)
}
\seealso{
Mukerjee, S. (2021). A systematic comparison of community detection algorithms for measuring selective exposure in co-exposure networks. Scientific Reports, 11(1), 1-11.
}
